analysis_service/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
EditorConfig==0.12.4
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
httpx-sse==0.4.0
idna==3.6
jiter==0.9.0
jsbeautifier==1.14.11
jsonpatch==1.33
jsonpointer==2.4
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.24
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.70.0
orjson==3.10.16
packaging==23.2
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pypdf==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.9.0
tqdm==4.66.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.2.0
uvicorn==0.27.1
yarl==1.9.4
zstandard==0.23.0